feat: restructure plugin architecture and add core components
