+++ src/main/java/ch/ldb/core/Observable.java
@@ -1,4 +1,4 @@
-package ch.ldb;
+package ch.ldb.core;
 
 import java.util.function.Consumer;
