--- src/main/java/ch/ldb/observable_demo/FormatPlugin.java
@@ -1,11 +1,0 @@
-package ch.ldb.observable_demo;
-
-public class FormatPlugin {
-    public String transformInput(String data) {
-        return "In: " + data;
-    }
-
-    public String transformOutput(String data) {
-        return "Out: " + data;
-    }
-}
