+++ src/main/java/ch/ldb/plugins/FilePlugin2.java
@@ -1,4 +1,4 @@
-package ch.ldb;
+package ch.ldb.plugins;
 
 import java.io.BufferedWriter;
 import java.io.FileWriter;
@@ -11,12 +11,16 @@ import java.nio.file.StandardWatchEventKinds;
 import java.nio.file.WatchEvent;
 import java.nio.file.WatchKey;
 import java.nio.file.WatchService;
+import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-public class FilePlugin1 implements Plugin<String> {
+import ch.ldb.core.Observable;
+
+public class FilePlugin2 implements Plugin<String> {
     private final Observable<String> output = new Observable<>();
     private final ExecutorService executor = Executors.newSingleThreadExecutor();
+    private long lastLineCount = 0; // Tracks the number of lines in the input file
     private String outputFilePath; // Path to the output file
 
     @Override
@@ -26,10 +30,10 @@ public class FilePlugin1 implements Plugin<String> {
 
     @Override
     public void setInput(Observable<String> input) {
-        // Subscribe to the input observable and overwrite the output file with the new data
+        // Subscribe to the input observable and append the data to the output file
         input.subscribe(data -> {
             if (outputFilePath != null) {
-                overwriteFile(outputFilePath, data); // Overwrite the output file
+                appendToFile(outputFilePath, data); // Append to the configured output file
             } else {
                 System.err.println("Output file path is not set!");
             }
@@ -52,8 +56,13 @@ public class FilePlugin1 implements Plugin<String> {
                     WatchKey key = watchService.take();
                     for (WatchEvent<?> event : key.pollEvents()) {
                         if (event.context().toString().equals(Paths.get(filePath).getFileName().toString())) {
-                            String content = Files.readString(Paths.get(filePath));
-                            output.emit(content); // Emit the entire file content
+                            List<String> lines = Files.readAllLines(Paths.get(filePath));
+                            if (lines.size() > lastLineCount) {
+                                // Get the newest line
+                                String newLine = lines.get(lines.size() - 1);
+                                lastLineCount = lines.size(); // Update the line count
+                                output.emit(newLine); // Emit the newest line
+                            }
                         }
                     }
                     key.reset();
@@ -68,9 +77,10 @@ public class FilePlugin1 implements Plugin<String> {
         this.outputFilePath = filePath; // Set the output file path
     }
 
-    private void overwriteFile(String filePath, String content) {
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(filePath, false))) { // Overwrite mode
+    private void appendToFile(String filePath, String content) {
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(filePath, true))) { // Append mode
             writer.write(content);
+            writer.newLine();
         } catch (IOException e) {
             e.printStackTrace();
         }
